modules/honeyenrich/src/requirements.txt
ipwhois==1.1.0
dnspython==1.16.0